enet: am263x/am263px/am261x: Fix CPTS adjust overflow in IOCTL
- Multiplying tsAdj->adjValInNsecs (int32_t) by -1 was performed using 32-bit arithmetic before casting.
  This risks undefined behavior and wrapping if the value is INT32_MIN.Fixed by casting to int64_t prior
  to negation to ensure safe 64-bit evaluation before assigning to adjOffset.

Fixes: MCUSDK-15794

+++ src/mod/cpsw_cpts_ioctl.c
@@ -281,7 +281,7 @@ int32_t CpswCpts_ioctl_handler_ENET_TIMESYNC_IOCTL_ADJUST_TIMESTAMP(CpswCpts_Han
                 ppmDir = CSL_CPTS_TS_PPM_DIR_DECREASE;
                 estfPpmDir = CSL_CPTS_ESTF_PPM_DIR_DECREASE;
                 genfPpmDir = CSL_CPTS_GENF_PPM_DIR_DECREASE;
-                adjOffset = (uint64_t)(-1 * tsAdj->adjValInNsecs);
+                adjOffset = (uint32_t)(-1 * (int64_t)tsAdj->adjValInNsecs);
             }
 
             adjVal = (uint64_t)tsAdj->intervalInNsecs / adjOffset;
